Remove SCHEDULED state
After discussions, we believe it would be simpler to have a status
limited to "available" or actively being used ("used").

In cases of visits being cancelled, we could move the state back
to available as we would have from scheduled.

This smaller enum set will help when calculating the balance in the
future as well, as we don't need to account for an additional state.