VB-5220 - Migrate a prisoner balance (#50)
* VB-5220 - Migrate a prisoner balance

Initial migration to onboard a prisoner from NOMIS to dps.

This PR includes a new service, handling migrating a balance
of both positive and negative values.

In the current system design, negative VOs are represented as their
own model with unique status and types.

Add integration tests to cover new migration logic.

* Add index on prisoner_id on new change_log table

* Remove SCHEDULED state

After discussions, we believe it would be simpler to have a status
limited to "available" or actively being used ("used").

In cases of visits being cancelled, we could move the state back
to available as we would have from scheduled.

This smaller enum set will help when calculating the balance in the
future as well, as we don't need to account for an additional state.

* VB-5220 - Initial migration

Add ability to track prisoners last allocated date regardless of their balance size.

This will allow us to simplify the allocation logic, to not care about the vo / negative_vo tables
and instead go directly to the prisoner_details table to get the last allocated date.

* make prisoner_id the primary key for new prisoner_details table

* clean up code layout and remove not needed unique on sql column

* move to use timestamp instead of date for vo tables

* compare dates to avoid timing issues

* change vo dates to use new prisoner details table - part1

* Update new prisoner details table to include a date per vo and pvo

Using these dates to now track allocation

* add cleanup before each test

* Attempt fix test with cleanup after each test

* Update as per PR comments