Update database and JPA entities to use relationships
Now we have a stable version of our initial database we can complete the
initial relationships between the prisoner_details table and the supporting
tables (visit_order, negative_visit_order and change_log).

A lot of small changes across the whole service to use the relationship
and pass in the prisoner entity into the supporting entities.