Move prisoner processing into separate class with managed transaction (#85)
* Move prisoner processing into separate class with managed transaction

Ready for when we need to publish event per prisoner after processing,
we need to ensure the prisoners transaction has fully committed before
raising the event.

* Return changeLog from processPrisoner to allow event publishing

In future the returned ChangeLog will be used to publish an
"adjustment" event, which allows NOMIS to call our service
and sync after we've changed a prisoners balance due to our
nightly batch job running.

* Rename column on prisoner audit job table to better describe function

now if a prisoner has no changes, we don't add a change log.

rename failed counter to failedOrSkipped