Bug fixes
Fix the possiblity that our SQS "at least once delivery" queue, can
provide the service with the same message twice if something goes
wrong on AWS side. Added a check on visit-booked event, to not
re-process the same message if another VO is already mapped to the visit.

Fixed some spacing issues