Bug - Fix transaction boundries causing locks to exit too early
As the top level service was outside of a transaction, the call
to get or create the new prisoner details, opened a transaction but
instantly closed after exiting the sub service. Leaving the top level
service to make the next call without a lock.

Issue spotted: This meant 2 transactions were racing to complete, and the 2nd
transaction found "orphaned" records from the 1st, and deleted them on transaction closure.

Fix: The transaction boundry should get or create the prisoner, and now keep the lock open and
make all changes / inserts before releasing.