Bug - Issue with Hibernate, JPA, data classes and orphanRemoval = true (#157)
* Bug - Issue with Hibernate, JPA, data classes and orphanRemoval = true

We believe a data issue is happening where requests are removing previous
requests saved child objects due to issues with using data classes and their default
equals / hashcode methods in kotlin. Hibernate seems to be marking the old objects are orphans
and cleaning them up incorrectly.

Fix: Add overrides for each of the equals / hashcode methods. Also remove the use of deepCopy
as this could cause issues in the future if a dev accidentally called to save this copied object.

* make classes open not data