Clean up JPA relationships removing duplicate FK prisoner_id column and let JPA manage it directly (#161)
* Clean up JPA relationships removing duplicate FK prisoner_id column and let JPA manage it directly

Instead of us specifying the column in the Entity, remove that and only have the ManyToOne on the prisonerDetails model field
which should let JPA manage it better.

* fix hibernate proxy warning by making fields open

* Clean up transactional annotation on change log service

It's possible this added annotation is causing the deletes on racing
requests. Removing it from the change log methods as the calling service
is managing the transaction.

* Correct allOpen to use jakarta

* move annotation onto it's own line