Bug fix: Cancelled visits not refunding VOs
Bug description:
If a negative visit order was assigned to a visit, and got repaid
via the nightly allocation. If the visit was then cancelled, the original
"remove element" functionality didn't repay the NVO, as it was already in a state
of repaid.

Fix description:
Improve the checking, if the NVO assigned to the visit, is already in a REPAID status,
try to repay another NVO, and as a final fallback, generate a new VO.