Prisoner merge events - change domain event raised after processing (#253)
Instead of an adjustment event, we raise a balance reset event. This is
needed, because we have no knowledge of the booking balances of each prisoner
on NOMIS's side. Instead of raising an adjustment, which can cause NOMIS to
fall out of sync.

The balance-reset event, forces NOMIS to call to get the current balance, and sync
to it, this fixes our issue.